* src/net/enet.c (ntohs): Return type is a short.         * src/flash.c (fis_list, fis_free, fis_create):         * src/iomem.c (do_iopeek): Fixed compiler warnings about formats         strings for diag_printf.

+++ packages/redboot/current/include/net/net.h
@@ -86,7 +86,7 @@ extern unsigned long get_ms_ticks(void);
 #define __LITTLE_ENDIAN__
 #endif
 extern unsigned long  ntohl(unsigned long x);
-extern unsigned long  ntohs(unsigned short x);
+extern unsigned short ntohs(unsigned short x);
 #else
 #define	ntohl(x)	(x)
 #define	ntohs(x)	(x)
